Change trim_gamepad_axis to return type double

+++ src/frontend/device.h
@@ -111,9 +111,9 @@ void override_joybus_devices_ptr(n64_joybus_device_t* override);
 
 // Exposed for testing
 
-// Trim and apply deadzone
-s8 trim_gamepad_axis(s16 raw);
-// do all requisite clamping for a controller
+// Trim input
+double trim_gamepad_axis(s16 raw);
+// Do all requisite clamping for a controller
 void clamp_gamepad(n64_controller_t* controller);
 
 #ifdef __cplusplus
